+++ frontend/src/views/EditView/User/MaaEndUserEdit.vue
@@ -65,6 +65,7 @@
             @save-batch="handleFieldsSave"
           />
           <SkylandConfigSection :form-data="formData" :loading="loading" @save="handleFieldSave" />
+          <ExtraScriptSection :form-data="formData" :loading="loading" @save="handleFieldSave" />
           <NotifyConfigSection
             :form-data="formData"
             :loading="loading"
@@ -95,6 +96,7 @@ import BasicInfoSection from '../../MaaEndUserEdit/BasicInfoSection.vue'
 import TaskConfigSection from '../../MaaEndUserEdit/TaskConfigSection.vue'
 import SkylandConfigSection from '../../MaaEndUserEdit/SkylandConfigSection.vue'
 import NotifyConfigSection from '../../MaaEndUserEdit/NotifyConfigSection.vue'
+import ExtraScriptSection from '@/components/ExtraScriptSection.vue'
 
 const logger = window.electronAPI.getLogger('MaaEnd用户编辑')
 
@@ -135,6 +137,10 @@ const getDefaultMaaEndUserData = () => ({
     SanityMode: 'Fixed',
     Resource: '官服',
     RemainedDay: -1,
+    IfScriptBeforeTask: false,
+    ScriptBeforeTask: '',
+    IfScriptAfterTask: false,
+    ScriptAfterTask: '',
     IfSkland: false,
     SklandToken: '',
     Notes: '',
